Create doubly_linked_list.c
Doubly Linked List Implementation in C

No compile error.
It implements a doubly linked list data structure, which is a type of data structure that contains a set of nodes with links to both the next and previous nodes in the list. This code demonstrates the creation of a new node, inserting a new node into the linked list, printing the linked list in both forward and reverse order, and freeing the linked list. The code also includes timing information for the overall execution time of the linked list operations.